refactor(events-handler): refactored event when change level
