+++ src/ApplicationServer/NeoServer.Server.Events/Player/PlayerLevelAdvancedEventHandler.cs
@@ -1,37 +1,16 @@
-﻿using NeoServer.Game.Common.Contracts.Creatures;
-using NeoServer.Game.Common.Creatures;
-using NeoServer.Game.Common.Helpers;
+﻿using NeoServer.Game.Common.Creatures;
 using NeoServer.Game.Common.Parsers;
 using NeoServer.Networking.Packets.Outgoing;
-using NeoServer.Networking.Packets.Outgoing.Player;
 using NeoServer.Server.Common.Contracts;
 using NeoServer.Server.Common.Contracts.Network;
 
 namespace NeoServer.Server.Events.Player;
 
-public class PlayerLevelAdvancedEventHandler
+public class PlayerLevelAdvancedEventHandler : PlayerLevelChangeEventHandler
 {
-    private readonly IGameServer game;
+    public PlayerLevelAdvancedEventHandler(IGameServer game) : base(game) { }
 
-    public PlayerLevelAdvancedEventHandler(IGameServer game)
-    {
-        this.game = game;
-    }
-
-    public void Execute(IPlayer player, SkillType skillType, int fromLevel, int toLevel)
-    {
-        if (Guard.IsNull(player)) return;
-
-        if (!game.CreatureManager.GetPlayerConnection(player.CreatureId, out var connection)) return;
-
-        SendLevelChangeMessage(skillType, connection, fromLevel, toLevel);
-
-        connection.OutgoingPackets.Enqueue(new PlayerStatusPacket(player));
-
-        connection.Send();
-    }
-
-    protected void SendLevelChangeMessage(SkillType skillType, IConnection connection, int fromLevel, int toLevel)
+    protected override void SendLevelChangeMessage(SkillType skillType, IConnection connection, int fromLevel, int toLevel)
     {
         connection.OutgoingPackets.Enqueue(new TextMessagePacket(
             MessageParser.GetSkillAdvancedMessage(skillType, fromLevel, toLevel),
