+++ src/ApplicationServer/NeoServer.Server.Events/Player/PlayerLevelRegressedEventHandler.cs
@@ -1,41 +1,19 @@
-﻿using NeoServer.Game.Common.Contracts.Creatures;
-using NeoServer.Game.Common.Creatures;
-using NeoServer.Game.Common.Helpers;
+﻿using NeoServer.Game.Common.Creatures;
 using NeoServer.Game.Common.Parsers;
-using NeoServer.Networking.Packets.Outgoing.Player;
 using NeoServer.Networking.Packets.Outgoing;
 using NeoServer.Server.Common.Contracts.Network;
 using NeoServer.Server.Common.Contracts;
 
-namespace NeoServer.Server.Events.Player
+namespace NeoServer.Server.Events.Player;
+
+public class PlayerLevelRegressedEventHandler : PlayerLevelChangeEventHandler
 {
-    public class PlayerLevelRegressedEventHandler
+    public PlayerLevelRegressedEventHandler(IGameServer game) : base(game) { }
+
+    protected override void SendLevelChangeMessage(SkillType skillType, IConnection connection, int fromLevel, int toLevel)
     {
-        private readonly IGameServer game;
-
-        public PlayerLevelRegressedEventHandler(IGameServer game)
-        {
-            this.game = game;
-        }
-
-        public void Execute(IPlayer player, SkillType skillType, int fromLevel, int toLevel)
-        {
-            if (Guard.IsNull(player)) return;
-
-            if (!game.CreatureManager.GetPlayerConnection(player.CreatureId, out var connection)) return;
-
-            SendLevelChangeMessage(skillType, connection, fromLevel, toLevel);
-
-            connection.OutgoingPackets.Enqueue(new PlayerStatusPacket(player));
-
-            connection.Send();
-        }
-
-        protected void SendLevelChangeMessage(SkillType skillType, IConnection connection, int fromLevel, int toLevel)
-        {
-            connection.OutgoingPackets.Enqueue(new TextMessagePacket(
-                MessageParser.GetSkillRegressedMessage(skillType, fromLevel, toLevel),
-                TextMessageOutgoingType.MESSAGE_EVENT_LEVEL_CHANGE));
-        }
+        connection.OutgoingPackets.Enqueue(new TextMessagePacket(
+            MessageParser.GetSkillRegressedMessage(skillType, fromLevel, toLevel),
+            TextMessageOutgoingType.MESSAGE_EVENT_LEVEL_CHANGE));
     }
 }
